requirements.txt
aiofiles==23.1.0
aiogram==2.6
aiohappyeyeballs==2.4.4
aiohttp==3.8.6
aioschedule==0.5.2
aiosignal==1.3.2
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==4.2
django-ckeditor==6.7.2
django-cuser==2017.3.16
django-filter==25.1
django-jasmine==0.5.1
django-jazzmin==2.6.0
django-js-asset==3.0.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
magic-filter==1.0.12
modeltranslation==0.25
multidict==6.1.0
openai==1.64.0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PyJWT==2.9.0
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
slugify==0.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
telebot==0.0.5
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3